Stub out interface, extract code
